extract dev.patrickgold.florisboard.lib.kotlin to org.florisboard.lib.kotlin
