+++ app/src/main/kotlin/dev/patrickgold/florisboard/ime/dictionary/UserDictionary.kt
@@ -38,7 +38,7 @@ import dev.patrickgold.florisboard.lib.FlorisLocale
 import dev.patrickgold.florisboard.lib.ValidationRule
 import dev.patrickgold.florisboard.lib.android.readText
 import dev.patrickgold.florisboard.lib.android.writeText
-import dev.patrickgold.florisboard.lib.kotlin.tryOrNull
+import org.florisboard.lib.kotlin.tryOrNull
 import java.lang.ref.WeakReference
 
 private const val WORDS_TABLE = "words"
