+++ app/src/main/kotlin/dev/patrickgold/florisboard/ime/editor/EditorContent.kt
@@ -16,7 +16,7 @@
 
 package dev.patrickgold.florisboard.ime.editor
 
-import dev.patrickgold.florisboard.lib.kotlin.safeSubstring
+import org.florisboard.lib.kotlin.safeSubstring
 
 /**
  * A snapshot window of an input editor content around the selection/cursor.
@@ -25,8 +25,8 @@ import dev.patrickgold.florisboard.lib.kotlin.safeSubstring
  * @property offset The offset of the whole editor content snapshot. `-1` indicates the value is unknown.
  * @property localSelection The selection reported by the editor, without [offset] included.
  * @property localComposing The composing region for the editor, without [offset] included.
- * @property localCurrentWord The current word for the editor (typically the same as [localComposing]), without
- *  [offset] included.
+ * @property localCurrentWord The current word for the editor (typically the same as [localComposing]), without [offset]
+ *     included.
  */
 data class EditorContent(
     val text: String,
