+++ app/src/main/kotlin/dev/patrickgold/florisboard/ime/keyboard/LayoutManager.kt
@@ -34,8 +34,6 @@ import dev.patrickgold.florisboard.lib.devtools.flogDebug
 import dev.patrickgold.florisboard.lib.devtools.flogWarning
 import dev.patrickgold.florisboard.lib.ext.ExtensionComponentName
 import dev.patrickgold.florisboard.lib.io.ZipUtils
-import dev.patrickgold.florisboard.lib.kotlin.DeferredResult
-import dev.patrickgold.florisboard.lib.kotlin.runCatchingAsync
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.Dispatchers
@@ -44,6 +42,8 @@ import kotlinx.coroutines.async
 import kotlinx.coroutines.cancel
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
+import org.florisboard.lib.kotlin.DeferredResult
+import org.florisboard.lib.kotlin.runCatchingAsync
 
 private data class LTN(
     val type: LayoutType,
