+++ app/src/main/kotlin/dev/patrickgold/florisboard/ime/nlp/BreakIteratorGroup.kt
@@ -18,9 +18,9 @@ package dev.patrickgold.florisboard.ime.nlp
 
 import android.icu.text.BreakIterator
 import dev.patrickgold.florisboard.lib.FlorisLocale
-import dev.patrickgold.florisboard.lib.kotlin.GuardedByLock
-import dev.patrickgold.florisboard.lib.kotlin.guardedByLock
 import io.github.reactivecircus.cache4k.Cache
+import org.florisboard.lib.kotlin.GuardedByLock
+import org.florisboard.lib.kotlin.guardedByLock
 import kotlin.contracts.InvocationKind
 import kotlin.contracts.contract
