+++ app/src/main/kotlin/dev/patrickgold/florisboard/lib/Validation.kt
@@ -20,8 +20,8 @@ import androidx.annotation.StringRes
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import dev.patrickgold.florisboard.lib.compose.stringRes
-import dev.patrickgold.florisboard.lib.kotlin.CurlyArg
-import dev.patrickgold.florisboard.lib.kotlin.curlyFormat
+import org.florisboard.lib.kotlin.CurlyArg
+import org.florisboard.lib.kotlin.curlyFormat
 import kotlin.contracts.contract
 import kotlin.reflect.KClass
