+++ app/src/main/kotlin/dev/patrickgold/florisboard/lib/android/AndroidSettings.kt
@@ -28,7 +28,7 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalLifecycleOwner
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleEventObserver
-import dev.patrickgold.florisboard.lib.kotlin.tryOrNull
+import org.florisboard.lib.kotlin.tryOrNull
 import java.lang.reflect.Modifier
 import kotlin.reflect.KClass
