+++ app/src/main/kotlin/dev/patrickgold/florisboard/lib/android/LaunchUtils.kt
@@ -24,7 +24,7 @@ import androidx.annotation.StringRes
 import dev.patrickgold.florisboard.R
 import dev.patrickgold.florisboard.lib.devtools.flogError
 import dev.patrickgold.florisboard.lib.io.FlorisRef
-import dev.patrickgold.florisboard.lib.kotlin.CurlyArg
+import org.florisboard.lib.kotlin.CurlyArg
 import kotlin.contracts.InvocationKind
 import kotlin.contracts.contract
 import kotlin.reflect.KClass
