+++ app/src/main/kotlin/dev/patrickgold/florisboard/lib/devtools/Devtools.kt
@@ -24,7 +24,7 @@ import dev.patrickgold.florisboard.BuildConfig
 import dev.patrickgold.florisboard.R
 import dev.patrickgold.florisboard.app.AppPrefs
 import dev.patrickgold.florisboard.lib.android.systemService
-import dev.patrickgold.florisboard.lib.kotlin.titlecase
+import dev.patrickgold.florisboard.lib.titlecase
 import dev.patrickgold.florisboard.lib.util.TimeUtils
 import dev.patrickgold.florisboard.lib.util.UnitUtils
 import java.io.BufferedReader
