+++ app/src/main/kotlin/dev/patrickgold/florisboard/lib/snygg/SnyggRule.kt
@@ -20,13 +20,13 @@ import androidx.compose.runtime.saveable.Saver
 import dev.patrickgold.florisboard.ime.input.InputShiftState
 import dev.patrickgold.florisboard.ime.text.key.KeyCode
 import dev.patrickgold.florisboard.ime.theme.FlorisImeUi
-import dev.patrickgold.florisboard.lib.kotlin.curlyFormat
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.descriptors.PrimitiveKind
 import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
 import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.encoding.Encoder
+import org.florisboard.lib.kotlin.curlyFormat
 
 @Serializable(with = SnyggRule.Serializer::class)
 data class SnyggRule(
