+++ build.gradle.kts
@@ -19,6 +19,7 @@ plugins {
     alias(libs.plugins.agp.library) apply false
     alias(libs.plugins.agp.test) apply false
     alias(libs.plugins.kotlin.android) apply false
+    alias(libs.plugins.kotlin.jvm) apply false
     alias(libs.plugins.kotlin.serialization) apply false
     alias(libs.plugins.ksp) apply false
     alias(libs.plugins.mannodermaus.android.junit5) apply false
