+++ lib/kotlin/src/main/kotlin/org/florisboard/lib/kotlin/Numbers.kt
@@ -14,18 +14,6 @@
  * limitations under the License.
  */
 
-package dev.patrickgold.florisboard.lib.kotlin
+package org.florisboard.lib.kotlin
 
-import kotlin.contracts.InvocationKind
-import kotlin.contracts.contract
-
-inline fun <R> tryOrNull(block: () -> R): R? {
-    contract {
-        callsInPlace(block, InvocationKind.EXACTLY_ONCE)
-    }
-    return try {
-        block()
-    } catch (_: Throwable) {
-        null
-    }
-}
+fun Number.toStringWithoutDotZero(): String = this.toString().removeSuffix(".0")
